requirements.txt
pyside6
pyinstaller
python-docx
docx2pdf
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-qt>=4.2.0
pytest-mock>=3.10.0
